+++ yalafi/handlers.py
@@ -161,3 +161,14 @@ def h_load_module(prefix):
         return utils.filter_set_toks(out, pos, None)
     return f
 
+
+# \MakeLowercase
+def h_makeLowercase(parser, buf, mac, args, delim, pos):
+    string = parser.get_text_expanded(args[0])
+    return [defs.TextToken(pos, string.lower())]
+
+
+# \MakeUppercase
+def h_makeUppercase(parser, buf, mac, args, delim, pos):
+    string = parser.get_text_expanded(args[0])
+    return [defs.TextToken(pos, string.upper())]